requirements.txt
# REMB Dependencies - Industrial Estate Master Planning Optimization Engine

# Core Optimization
pymoo>=0.6.0
ortools>=9.7.2876

# Geometry Processing
geopandas>=0.14.0
shapely>=2.0.0
ezdxf>=1.1.0
pyproj>=3.6.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
geoalchemy2>=0.14.0
alembic>=1.12.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
